+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 - Refactored `image_quality.get_default_phenocam_flags` to `data_products.phenocams.get_default_phenocam_flags`
 - Refactored `image_quality.load_flags_weights` to `data_products.phenocams.load_flags_weights`
 - Code-pruning `station.populate_station_db` and `station.create_record_dictionary`
+- Improved `utils.mean_datetime_str` to check whether is a datetime onject or a formated datetime string
 
 ## [v0.14.0]
 -  updated `version_schema_platform_phenocams='0.2-2024'`